28/12/2022

+++ bundle/models/product_bundle_qty.py
@@ -5,8 +5,8 @@ class ProductBundle(models.Model):
     _name = "product.bundle.qty"
 
     is_add_range = fields.Boolean()
-    qty_start = fields.Integer()
-    qty_end = fields.Integer()
+    qty_start = fields.Float()
+    qty_end = fields.Float()
     discount_value = fields.Float()
     bundle_id = fields.Many2one('product.bundle')
